requirements.txt
annotated-types==0.7.0
certifi==2026.2.25
charset-normalizer==3.4.5
greenlet==3.3.2
idna==3.11
numpy==2.4.2
pandas==3.0.1
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
pyodbc==5.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
requests==2.32.5
six==1.17.0
SQLAlchemy==2.0.48
tenacity==9.1.4
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
